requirements.txt
python-dotenv
pymysql
streamlit
altair
pandas
boto3
sqlalchemy
pylint
pytest